Eliminación botón Login del header
