--- 265471.txt
@@ -1,3 +1,0 @@
-271616
-267771
-284075
