+++ PolideportivoUsuario_Proj/PolideportivoUsuario_Proj.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PolideportivoUsuario_Proj", "Polideportivo_Proj\PolideportivoUsuario_Proj.csproj", "{44C1719F-2B13-4FFC-B527-DED197A1F800}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{44C1719F-2B13-4FFC-B527-DED197A1F800}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44C1719F-2B13-4FFC-B527-DED197A1F800}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44C1719F-2B13-4FFC-B527-DED197A1F800}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44C1719F-2B13-4FFC-B527-DED197A1F800}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CA2A7014-F561-46E6-B3EF-8CEDC3CDAD21}
+	EndGlobalSection
+EndGlobal
