--- Polideportivo_Proj/Polideportivo_Proj/Mantenimientos/Admin. Financiera/FrmAdminFinanc.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Polideportivo_Proj.Mantenimientos.Admin._Financiera
-{
-    public partial class FrmAdminFinanc : Form
-    {
-        public FrmAdminFinanc()
-        {
-            InitializeComponent();
-        }
-    }
-}
