--- Polideportivo_Proj/Polideportivo_Proj/Mantenimientos/Main/FrmMain.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Polideportivo_Proj
-{
-    public partial class FrmMain : Form
-    {
-        public FrmMain()
-        {
-            InitializeComponent();
-        }
-
-        private void BtnListadoE_Click(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
